# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Climate Change Visualization
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TemperatureRecord {
  year: number;
  region: string;
  /** Average temperature in Celsius */
  temperature: number;
  /** Temperature anomaly vs 1981-2010 baseline in Celsius */
  temperatureAnomaly: number;
  scenario: string;
  model: string;
}

export interface DataMetadata {
  source: string;
  model: string;
  baseline: string;
  unit: string;
}

export interface TemperatureDataResponse {
  data: TemperatureRecord[];
  metadata: DataMetadata;
}

export interface HumidityRecord {
  year: number;
  region: string;
  /** Relative humidity percentage */
  humidity: number;
  /** Specific humidity in g/kg */
  specificHumidity: number;
  scenario: string;
}

export interface HumidityDataResponse {
  data: HumidityRecord[];
  metadata: DataMetadata;
}

export interface GlobePoint {
  lat: number;
  lon: number;
  temperature: number;
  temperatureAnomaly: number;
}

export interface GlobeDataResponse {
  year: number;
  scenario: string;
  points: GlobePoint[];
}

export interface Region {
  id: string;
  name: string;
  lat: number;
  lon: number;
  color: string;
}

export interface RegionsResponse {
  regions: Region[];
}

export interface Scenario {
  id: string;
  name: string;
  description: string;
  warming: string;
}

export interface ScenariosResponse {
  scenarios: Scenario[];
}

export type GetTemperatureDataParams = {
  /**
   * Filter by region (e.g., Africa, Asia, Europe)
   */
  region?: string;
  /**
   * Climate scenario (SSP pathway)
   */
  scenario?: GetTemperatureDataScenario;
};

export type GetTemperatureDataScenario =
  (typeof GetTemperatureDataScenario)[keyof typeof GetTemperatureDataScenario];

export const GetTemperatureDataScenario = {
  ssp126: "ssp126",
  ssp245: "ssp245",
  ssp370: "ssp370",
  ssp585: "ssp585",
} as const;

export type GetHumidityDataParams = {
  /**
   * Filter by year
   */
  year?: number;
  /**
   * Climate scenario
   */
  scenario?: GetHumidityDataScenario;
};

export type GetHumidityDataScenario =
  (typeof GetHumidityDataScenario)[keyof typeof GetHumidityDataScenario];

export const GetHumidityDataScenario = {
  ssp126: "ssp126",
  ssp245: "ssp245",
  ssp370: "ssp370",
  ssp585: "ssp585",
} as const;

export type GetGlobeDataParams = {
  /**
   * Year for globe data
   */
  year?: number;
  scenario?: GetGlobeDataScenario;
};

export type GetGlobeDataScenario =
  (typeof GetGlobeDataScenario)[keyof typeof GetGlobeDataScenario];

export const GetGlobeDataScenario = {
  ssp126: "ssp126",
  ssp245: "ssp245",
  ssp370: "ssp370",
  ssp585: "ssp585",
} as const;