# lib/api-zod/src/generated/types/globeDataResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Climate Change Visualization
 * OpenAPI spec version: 0.1.0
 */
import type { GlobePoint } from "./globePoint";

export interface GlobeDataResponse {
  year: number;
  scenario: string;
  points: GlobePoint[];
}