# lib/api-zod/src/generated/types/globePoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Climate Change Visualization
 * OpenAPI spec version: 0.1.0
 */

export interface GlobePoint {
  lat: number;
  lon: number;
  temperature: number;
  temperatureAnomaly: number;
}