# lib/api-zod/src/generated/types/humidityRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Climate Change Visualization
 * OpenAPI spec version: 0.1.0
 */

export interface HumidityRecord {
  year: number;
  region: string;
  /** Relative humidity percentage */
  humidity: number;
  /** Specific humidity in g/kg */
  specificHumidity: number;
  scenario: string;
}