# lib/api-zod/src/generated/types/regionsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Climate Change Visualization
 * OpenAPI spec version: 0.1.0
 */
import type { Region } from "./region";

export interface RegionsResponse {
  regions: Region[];
}