# lib/api-zod/src/generated/types/scenario.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Climate Change Visualization
 * OpenAPI spec version: 0.1.0
 */

export interface Scenario {
  id: string;
  name: string;
  description: string;
  warming: string;
}