# lib/api-zod/src/generated/types/temperatureDataResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Climate Change Visualization
 * OpenAPI spec version: 0.1.0
 */
import type { DataMetadata } from "./dataMetadata";
import type { TemperatureRecord } from "./temperatureRecord";

export interface TemperatureDataResponse {
  data: TemperatureRecord[];
  metadata: DataMetadata;
}